refactor: change delete community emoji api (#103)
